🧪 Create dedicated test file for binomial_theorem.py
Extracted tests for `binomial_theorem.py` out of `test_DiscreteMath.py` into a focused, dedicated test file `Tests/test_binomial_theorem.py` to improve project structure.
